+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 9.3
 - Use DBLastInsertID() instead of DBSeqNextID() in Moodle/includes/ImportUsers.fnc.php
 - Still use DBSeqNextID() for student ID, adapt for MySQL in Student.php & Moodle/includes/ImportUsers.fnc.php
 - SQL use CONCAT() instead of pipes || for MySQL compatibility, program wide
+- Fix first item in the list not displayed in Accounting/includes/DailyTransactions.php
 
 Changes in 9.2.2
 ----------------